error_file and out_file paths
In attempting to set these values, I assumed that the pathnames were relative to the ~/.pm2 directory. Oddly, if I leave the path off altogether, my log file is placed relative to the ~/.pm/ directory, however, if I specify a path (either ./logs/pm2.log or ~/.pm2/logs/pm2.log) it fails saying that it cannot find that directory. The only way I was able to make this work was to specify an absolute path form the base.